+++ src/components/BreakApp_/BreakApp.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { PokemonContext } from '../Context/PokemonContextProvider';
+import { PokemonContext } from '../Context_/PokemonContextProvider';
 
 const BreakApp = () => {
   const { setIsBreak } = useContext(PokemonContext) || {};
