+++ src/components/ErrorBoundary_/_errorBoundary.scss
