+++ src/components/ItemPerPage_/ItemPerPage.tsx
@@ -1,5 +1,5 @@
 import { useContext, useMemo } from 'react';
-import { PokemonContext } from '../Context/PokemonContextProvider';
+import { PokemonContext } from '../Context_/PokemonContextProvider';
 
 const ItemPerPage = () => {
   const { setSearchParams, setPokemonPerPage, setCurrentPage, setOffset } =
