+++ src/components/MainPage_/_mainPage.scss
