+++ src/components/NotFound_/404.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import App from '../App/App';
+import App from '../App_/App';
 import { MemoryRouter } from 'react-router-dom';
 
 describe('Tests for the 404 Page component:', () => {
