+++ src/components/Pokemon/PokemonBaseInfo/PokemonBaseInfo.tsx
