+++ src/components/Pokemon/PokemonCardAdditional/PokemonCardAdditional.test.tsx
@@ -3,7 +3,7 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { IPokemon } from '../../../types';
 import PokemonCardAdditional from './PokemonCardAdditional';
-import App from '../../App/App';
+import App from '../../App_/App';
 
 describe('Tests for the Detailed Card component:', () => {
   const pokemonObj: IPokemon = {
