+++ src/components/Pokemon/_pokemonList.scss
